refactor(utils): move the pack and unpack proof functions to the utils package
