+++ packages/eddsa-proof/src/generate.ts
@@ -2,9 +2,9 @@ import { BigNumber } from "@ethersproject/bignumber"
 import { BytesLike, Hexable } from "@ethersproject/bytes"
 import { deriveSecretScalar } from "@zk-kit/eddsa-poseidon"
 import { NumericString, groth16 } from "snarkjs"
+import { packProof } from "@zk-kit/utils"
 import getSnarkArtifacts from "./get-snark-artifacts.node"
 import hash from "./hash"
-import packProof from "./pack-proof"
 import { EddsaProof, SnarkArtifacts } from "./types"
 
 /**
