--- packages/eddsa-proof/src/pack-proof.ts
@@ -1,20 +1,0 @@
-import { Groth16Proof } from "@zk-kit/groth16"
-import { PackedProof } from "./types"
-
-/**
- * Packs a proof into a format compatible with EddsaProof.
- * @param proof The Groth16 proof generated with SnarkJS.
- * @returns The proof compatible with EddsaProof.
- */
-export default function packProof(proof: Groth16Proof): PackedProof {
-    return [
-        proof.pi_a[0],
-        proof.pi_a[1],
-        proof.pi_b[0][1],
-        proof.pi_b[0][0],
-        proof.pi_b[1][1],
-        proof.pi_b[1][0],
-        proof.pi_c[0],
-        proof.pi_c[1]
-    ]
-}
