+++ packages/poseidon-proof/src/generate.ts
@@ -1,9 +1,9 @@
 import { BigNumber } from "@ethersproject/bignumber"
 import { BytesLike, Hexable } from "@ethersproject/bytes"
 import { NumericString, groth16 } from "snarkjs"
+import { packProof } from "@zk-kit/utils"
 import getSnarkArtifacts from "./get-snark-artifacts.node"
 import hash from "./hash"
-import packProof from "./pack-proof"
 import { PoseidonProof, SnarkArtifacts } from "./types"
 
 /**
