+++ packages/utils/src/index.ts
@@ -5,3 +5,4 @@ export { F1Field, scalar }
 export * from "./types"
 export * from "./number-checks"
 export * from "./number-conversions"
+export * from "./groth16-proof"
